+++ src/components/MediaModal.tsx
@@ -118,8 +118,6 @@ const MediaModal = ({
         "transform",
         `translate3d(0px, ${x}px,${0}px)`
       );
-      translateDiv.current.dataset.trodditMotionAt = String(performance.now());
-      translateDiv.current.dataset.trodditWriterHeight = String(windowHeight);
     }
   };
   // const [animateY, setAnimateY] = useState(false);
@@ -364,9 +362,6 @@ const MediaModal = ({
       )}
       <div
         ref={translateDiv}
-        data-troddit-media-rail
-        data-troddit-index={curPostNum}
-        data-troddit-zoom={scale}
         className={"absolute top-0 left-0 "}
         onTouchStart={(e) =>
           !flattenedPosts[curPostNum]?.data?.mediaInfo?.isSelf &&